dl_project/requirements.txt
click
opencv-python
torch
torchvision
configparser
Pillow
numpy
scipy
matplotlib